requirements.txt
netaddr==0.7.19
pyTenable>=0.3.26
ubiltools>=1.0.4
tenable-sc-config>=1.0.2